chore(chart): use getLinesMinMax
