+++ packages/chart/src/utils/constants.ts
@@ -1,3 +1,13 @@
 export const chartPaddingHorizontal = 40;
 export const chartPaddingVertical = 20;
 export const defaultChartHeightDivider = 4;
+export const COLORS = [
+  '#E0F5E9',
+  '#9BE1DA',
+  '#56CDCB',
+  '#3B8B8E',
+  '#EF1E1C',
+  '#F43B1D',
+  '#F9571D',
+  '#FF8A57',
+];
